add #[panic_handler]; deprecate #[panic_implementation]
