+++ src/test/run-make/wasm-symbols-not-exported/bar.rs
@@ -8,7 +8,7 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-#![feature(panic_implementation, alloc_error_handler)]
+#![feature(panic_handler, alloc_error_handler)]
 #![crate_type = "cdylib"]
 #![no_std]
 
@@ -39,7 +39,7 @@ fn a(_: core::alloc::Layout) -> ! {
     loop {}
 }
 
-#[panic_implementation]
+#[panic_handler]
 fn b(_: &core::panic::PanicInfo) -> ! {
     loop {}
 }
