+++ src/test/ui/feature-gates/feature-gate-panic-handler.stderr
@@ -1,0 +1,11 @@
+error[E0658]: #[panic_handler] is an unstable feature (see issue #44489)
+  --> $DIR/feature-gate-panic-handler.rs:18:1
+   |
+LL | #[panic_handler] //~ ERROR #[panic_handler] is an unstable feature (see issue #44489)
+   | ^^^^^^^^^^^^^^^^
+   |
+   = help: add #![feature(panic_handler)] to the crate attributes to enable
+
+error: aborting due to previous error
+
+For more information about this error, try `rustc --explain E0658`.
