+++ src/test/ui/missing/missing-alloc_error_handler.rs
@@ -13,9 +13,9 @@
 
 #![no_std]
 #![crate_type = "staticlib"]
-#![feature(panic_implementation, alloc_error_handler, alloc)]
+#![feature(panic_handler, alloc_error_handler, alloc)]
 
-#[panic_implementation]
+#[panic_handler]
 fn panic(_: &core::panic::PanicInfo) -> ! {
     loop {}
 }
