+++ src/test/ui/panic-handler/panic-handler-bad-signature-1.stderr
@@ -1,11 +1,11 @@
 error: return type should be `!`
-  --> $DIR/panic-implementation-bad-signature-1.rs:22:6
+  --> $DIR/panic-handler-bad-signature-1.rs:22:6
    |
 LL | ) -> () //~ ERROR return type should be `!`
    |      ^^
 
 error: argument should be `&PanicInfo`
-  --> $DIR/panic-implementation-bad-signature-1.rs:21:11
+  --> $DIR/panic-handler-bad-signature-1.rs:21:11
    |
 LL |     info: PanicInfo, //~ ERROR argument should be `&PanicInfo`
    |           ^^^^^^^^^
