+++ src/test/ui/panic-handler/panic-handler-bad-signature-2.stderr
@@ -1,5 +1,5 @@
 error: argument should be `&PanicInfo`
-  --> $DIR/panic-implementation-bad-signature-2.rs:21:11
+  --> $DIR/panic-handler-bad-signature-2.rs:21:11
    |
 LL |     info: &'static PanicInfo, //~ ERROR argument should be `&PanicInfo`
    |           ^^^^^^^^^^^^^^^^^^
