+++ src/test/ui/panic-handler/panic-handler-bad-signature-3.rs
@@ -10,19 +10,13 @@
 
 // compile-flags:-C panic=abort
 
-#![feature(lang_items)]
-#![feature(panic_implementation)]
+#![feature(panic_handler)]
 #![no_std]
 #![no_main]
 
 use core::panic::PanicInfo;
 
-#[panic_implementation]
-fn panic(info: &PanicInfo) -> ! {
-    loop {}
-}
-
-#[lang = "panic_impl"]
-fn panic2(info: &PanicInfo) -> ! { //~ ERROR duplicate lang item found: `panic_impl`.
+#[panic_handler]
+fn panic() -> ! { //~ ERROR function should have one argument
     loop {}
 }
