+++ src/test/ui/panic-handler/panic-handler-requires-panic-info.stderr
