+++ src/test/ui/panic-handler/panic-handler-std.stderr
@@ -1,5 +1,5 @@
 error[E0152]: duplicate lang item found: `panic_impl`.
-  --> $DIR/panic-implementation-std.rs:18:1
+  --> $DIR/panic-handler-std.rs:18:1
    |
 LL | / fn panic(info: PanicInfo) -> ! {
 LL | |     loop {}
