--- src/test/ui/panic-implementation/panic-implementation-bad-signature-4.stderr
@@ -1,11 +1,0 @@
-error: `#[panic_implementation]` function should have no type parameters
-  --> $DIR/panic-implementation-bad-signature-4.rs:20:1
-   |
-LL | / fn panic<T>(pi: &PanicInfo) -> ! {
-LL | |     //~^ ERROR `#[panic_implementation]` function should have no type parameters
-LL | |     loop {}
-LL | | }
-   | |_^
-
-error: aborting due to previous error
-
