+++ src/test/ui/panic-implementation/panic-implementation-deprecated.stderr
@@ -1,0 +1,14 @@
+error: use of deprecated attribute `panic_implementation`: This attribute was renamed to `panic_handler`. See https://github.com/rust-lang/rust/issues/44489#issuecomment-415140224
+  --> $DIR/panic-implementation-deprecated.rs:19:1
+   |
+LL | #[panic_implementation]
+   | ^^^^^^^^^^^^^^^^^^^^^^^ help: remove this attribute
+   |
+note: lint level defined here
+  --> $DIR/panic-implementation-deprecated.rs:13:9
+   |
+LL | #![deny(deprecated)]
+   |         ^^^^^^^^^^
+
+error: aborting due to previous error
+
