+++ src/test/ui/panic_implementation-closures.rs
@@ -12,9 +12,9 @@
 
 #![crate_type = "rlib"]
 #![no_std]
-#![feature(panic_implementation)]
+#![feature(panic_handler)]
 
-#[panic_implementation]
+#[panic_handler]
 pub fn panic_fmt(_: &::core::panic::PanicInfo) -> ! {
     |x: u8| x;
     loop {}
